Compute curDir relatively to allow subdirectories
The previous pattern of splitting the directory name on the docs path is
analogous to breaking up "/path/to/repo/docs/subdir/file.md" into two
pieces: "/path/to/repo" and "subdir/file.md" (assuming we're using the
default docs directory of "docs").

This doesn't account for paths that might have successive "docs"
subdirectories, such as a tree that looks like:

```
repo/
  docs/
    projectA/
      README.md
      docs/
        ARCHITECTURE.md
```

In this tree, README.md will be found and calculated correctly (relative
to projectA/) but ARCHITECTURE.md will not be found because it there is
a second "docs" directory in the chain, causing .split() to result in 3
elements, not 2 as expected.

The new pattern shown here calculates the subdirectory (`curDir`) of the
current file relative to the docs folder that is being processed.